feat: updated backend for ownership

+++ server/routers/routes.js
@@ -1,12 +1,23 @@
 const express = require('express');
 const Quiz = require('../models/model');
+const Questions = require('../models/questionModel');
+const User = require('../models/user')
 const router = express.Router();
 
 // Post Method - Create a new quiz
-router.post('/quiz', async(req, res) => {
+router.post('/quiz/:id', async(req, res) => {
     try {
+        const userId = req.params.id;
         const quiz = new Quiz(req.body);
+        const user = await User.findById(userId)
+        if (!user) {
+            return res.status(400).send({
+                error: 'no such user exists to add quiz'
+            })
+        }
         await quiz.save();
+        user.quizzes = user.quizzes.concat(quiz._id)
+        await user.save();
         res.status(201).send(quiz);
     } catch (error) {
         res.status(400).send(error);
@@ -60,11 +71,26 @@ router.patch('/quiz/:id', async(req, res) => {
 // Delete quiz by ID
 router.delete('/quiz/:id', async(req, res) => {
     try {
+        const { userId } = req.body
+        const user = await User.findById(userId)
+        if (!user) {
+            return res.status(400).send({
+                error: 'no such user exists to remove the quiz from'
+            })
+        }
+        // Removing questions that belongs to that quiz
+        const questionsIds = (await Quiz.findById(req.params.id)).questions.map(id => id.toJSON())
+        await Promise.all(questionsIds.map(id => Questions.findByIdAndDelete(id)))
+            // Removing Quiz
         const quiz = await Quiz.findByIdAndDelete(req.params.id);
         if (!quiz) {
             return res.status(404).send();
         }
-        res.send(quiz);
+        // Update the user
+        user.quizzes = user.quizzes.filter(id => id.toJSON() !== req.params.id)
+        await user.save()
+
+        res.status(204);
     } catch (error) {
         res.status(500).send(error);
     }
